Move rebin error propagation function to utils.
